feat: update User model to use Long for userId; refactor UserService and MapperTests accordingly

+++ api/src/main/java/dev/kalbarczyk/api/core/user/User.java
@@ -5,7 +5,7 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
 
 public record User(
-        int userId,
+        Long userId,
         @NotBlank @Size(min = 3, max = 50) String username,
         @NotBlank @Email String email,
         @NotBlank @Size(min = 3, max = 100) String password,
